split TransactionDUnit to reduce running time
- use common servers for tests in TransactionDUnit instead of starting/stopping
  in every test
- moved out tests that need special node settings (server-groups) into separate
  Transaction2DUnit; added extensions for RR and OffHeap like for TransactionDUnit

Together these two reduce running time from >25mins to ~3+2mins

- fixed table drop in JDBC.java to honour colocation chain
- dynamically update log-level in TestUtil method for cases where node is not being
  restarted at the end of test
