+++ gemfirexd/tools/src/dunit/java/com/pivotal/gemfirexd/internal/transactions/OffHeapTransaction2DUnit.java
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2017 SnappyData, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You
+ * may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License. See accompanying
+ * LICENSE file.
+ */
+
+package com.pivotal.gemfirexd.internal.transactions;
+
+public class OffHeapTransaction2DUnit extends Transaction2DUnit {
+
+  public OffHeapTransaction2DUnit(String name) {
+    super(name);
+  }
+
+  @Override
+  public void setUp() throws Exception {
+    super.setUp();
+    this.configureDefaultOffHeap(true);
+  }
+
+  @Override
+  public String getSuffix() {
+    return " offheap";
+  }
+}
