+++ gradle.properties
@@ -1,5 +1,5 @@
 org.gradle.daemon = false
-org.gradle.parallel=true
+#org.gradle.parallel=true
 
 # Set this on the command line with -P or in ~/.gradle/gradle.properties
 # to change the buildDir location.  Use an absolute path.
